AI-AGENT/requirements.txt
langchain 
langchain-community 
langchain-google-genai 
python-dotenv
pydantic
duckduckgo-search
beautifulsoup4